text2sql-server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlglot>=26.0.0
pyyaml>=6.0
httpx>=0.28.0
python-multipart>=0.0.20
click>=8.1.0
rich>=13.0.0
